Fixed latitude and longitude type

+++ ChatService/ChatService/Entities/Message.cs
@@ -8,8 +8,8 @@ namespace ChatService.Entities
         public string Author { get; set; }
         public string Country { get; set; }
         public string City { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
         public string Text { get; set; }
         public DateTime CreatedDate { get; set; }
     }
